pyspark-snippets/requirements.txt
pyspark==3.2.1
delta-spark==1.1.0
